Use the same structure for the submenu of Members that we used in other modules
